requirements.txt
absl-py==1.4.0
accelerate==0.25.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asttokens==2.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.12.2
bitsandbytes==0.41.3.post2
blinker==1.7.0
bs4==0.0.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
configparser==6.0.0
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.0
decorator==5.1.1
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.13.1
flask==3.0.0
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.0
fsspec==2023.12.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.62.0
greenlet==2.0.2
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.20.1
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
imutils==0.5.4
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter-client==8.3.1
jupyter-core==5.3.2
keras==2.15.0
kiwisolver==1.4.5
langchain==0.0.295
langsmith==0.0.38
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mediapipe==0.10.5
ml-dtypes==0.2.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
numexpr==2.8.6
numpy==1.26.3
oauthlib==3.2.2
openai==0.28.0
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.11.0
plotly==5.18.0
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.3.0
pydantic-core==2.6.3
PyGetWindow==0.0.9
Pygments==2.16.1
PyMsgBox==1.0.9
pyodbc==5.0.1
pyparsing==3.1.1
pypdf2==3.0.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pyspark==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
selenium==4.16.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.5
SQLAlchemy==2.0.21
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.36.2
trio==0.23.2
trio-websocket==0.11.1
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.4
ultralytics==8.1.2
urllib3==2.1.0
wcwidth==0.2.8
werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
